remove legacy preview loader
