+++ src/ActionMonitor/Monitors/PreviewMonitor.php
@@ -4,43 +4,23 @@ namespace WPGatsby\ActionMonitor\Monitors;
 
 use GraphQLRelay\Relay;
 use WPGatsby\Admin\Preview;
+use WPGatsby\Admin\Settings;
 
 class PreviewMonitor extends Monitor {
-	public static $last_sent_modified_time_key = '_wpgatsby_last_preview_modified_time';
-
 	function init() {
-		$enable_gatsby_preview = self::get_setting( 'enable_gatsby_preview' );
+		$enable_gatsby_preview = Settings::get_setting( 'enable_gatsby_preview' ) === 'on';
 
-		if ( $enable_gatsby_preview === 'on' ) {
+		if ( $enable_gatsby_preview ) {
 			add_filter( 'template_include', [ $this, 'setup_preview_template' ], 1, 99 );
 
-			add_action(
-				'graphql_register_types',
-				function() {
-					$this->registerPreviewStatusFieldsAndMutations();
-				}
-			);
-
-
-			$use_cloud_loader = self::get_setting( 'use_gatsby_content_sync' );
-
-			if ( $use_cloud_loader !== "on" ) {
-				add_action(
-					'save_post',
-					[ $this, 'post_to_preview_instance' ],
-					10,
-					2
-				);
-			}
-
-			add_filter( 'preview_post_link', function( $link, $post ) use ( $use_cloud_loader ) {
+			add_filter( 'preview_post_link', function( $link, $post ) {
 				$doing_graphql_request
 					= defined( 'GRAPHQL_REQUEST' ) && true === GRAPHQL_REQUEST;
 
 				// Use the normal $link during graphql requests
 				// because otherwise we could override the post URI
 				// in Gatsby! meaning the content sync or preview url could be added to the page path in Gatsby if pages are created from the uri.
-				if ( $doing_graphql_request || $use_cloud_loader !== "on") {
+				if ( $doing_graphql_request ) {
 					return $link;
 				}
 
@@ -49,511 +29,26 @@ class PreviewMonitor extends Monitor {
 		}
 	}
 
-	public static function get_gatsby_content_sync_url_for_post( $post ) {
-		// get the Gatsby Cloud loader url w/ site id
-		$gatsby_content_sync_url = self::get_setting( 'gatsby_content_sync_url' );
-					
-		// create the dynamic path the loader will need
-		$manifest_id = Preview::get_preview_manifest_id_for_post( $post );
-		$path = "/gatsby-source-wordpress/$manifest_id";
-
-		$url = preg_replace(
-			// remove any double forward slashes from the path
-			'/([^:])(\/{2,})/', '$1/',
-			"$gatsby_content_sync_url$path"
-		);
-
-		return $url;
-	}
-
-	public static function print_initial_preview_template_state_js() {
-		global $post;
-		$post_id = $post->ID;
-
-		$preview_url = self::get_gatsby_preview_instance_url();
-		$preview_url = rtrim( $preview_url, '/' );
-
-		$preview_webhook_online = get_option(
-			'_wp_gatsby_preview_webhook_is_online'
-		);
-
-		$initial_state = json_encode(
-			[
-				'postId'                 => $post_id,
-				'previewFrontendUrl'     => $preview_url,
-				'previewWebhookIsOnline' => $preview_webhook_online,
-				'graphqlEndpoint'        => Router::$route,
-				'webhookWasCalled'       => Preview::was_preview_webhook_called_for_post_id(
-					$post_id
-				),
-				'wordpressSiteUrl' 		=> get_site_url()
-			]
-		);
-
-		echo "var initialState = $initial_state; console.log({ initialState: initialState });";
-	}
-
-	function registerPreviewStatusFieldsAndMutations() {
-		register_graphql_enum_type(
-			'WPGatsbyRemotePreviewStatusEnum',
-			[
-				'description' => __( 'The different statuses a Gatsby Preview can be in for a single node.', 'wp-gatsby' ),
-				'values'      => [
-					'PREVIEW_SUCCESS'                      => [
-						'value' => 'PREVIEW_SUCCESS',
-					],
-					'NO_PAGE_CREATED_FOR_PREVIEWED_NODE'   => [
-						'value' => 'NO_PAGE_CREATED_FOR_PREVIEWED_NODE',
-					],
-					'GATSBY_PREVIEW_PROCESS_ERROR'         => [
-						'value' => 'GATSBY_PREVIEW_PROCESS_ERROR',
-					],
-					'RECEIVED_PREVIEW_DATA_FROM_WRONG_URL' => [
-						'value' => 'RECEIVED_PREVIEW_DATA_FROM_WRONG_URL',
-					],
-				],
-			]
-		);
-
-		register_graphql_mutation(
-			'wpGatsbyRemotePreviewStatus',
-			[
-				'inputFields'         => [
-					'pagePath'         => [
-						'type'        => 'String',
-						'description' => __( 'The Gatsby page path for this preview.', 'wp-gatsby' ),
-					],
-					'modified'         => [
-						'type'        => 'String',
-						'description' => __( 'The modified date of the latest revision for this preview.', 'wp-gatsby' ),
-					],
-					'parentDatabaseId' => [
-						'type'        => 'Number',
-						'description' => __( 'The previewed revisions post parent id', 'wp-gatsby' ),
-					],
-					'status'           => [
-						'type'        => [ 'non_null' => 'WPGatsbyRemotePreviewStatusEnum' ],
-						'description' => __( 'The remote status of the previewed node', 'wp-gatsby' ),
-					],
-					'statusContext'    => [
-						'type'        => 'String',
-						'description' => __( 'Additional context about the preview status', 'wp-gatsby' ),
-					],
-				],
-				'outputFields'        => [
-					'success' => [
-						'type'        => 'Boolean',
-						'description' => __( 'Wether or not the revision mutation was successful', 'wp-gatsby' ),
-						'resolve'     => function( $payload, $args, $context, $info ) {
-							$success = $payload['success'] ?? null;
-
-							return [
-								'success' => $success,
-							];
-						},
-					],
-				],
-				'mutateAndGetPayload' => function( $input, $context, $info ) {
-					$page_path       = $input['pagePath'] ?? null;
-					$modified        = $input['modified'] ?? null;
-					$parent_id       = $input['parentDatabaseId'] ?? null;
-					$remote_status   = $input['status'] ?? null;
-					$preview_context = $input['statusContext'] ?? null;
-
-					$post = get_post( $parent_id );
-
-					$post_type_object = $post
-						? get_post_type_object( $post->post_type )
-						: null;
-
-					$user_can_edit_this_post = $post
-						? current_user_can(
-							$post_type_object->cap->edit_posts,
-							$parent_id
-						)
-						: null;
-
-					if ( ! $post || ! $user_can_edit_this_post ) {
-						$message = sprintf(
-							__(
-								'Sorry, you are not allowed to update post %1$s',
-								'wp-gatsby'
-							),
-							$parent_id
-						);
-
-						throw new UserError( $message );
-					}
-
-					if ( $page_path ) {
-						update_post_meta(
-							$parent_id,
-							'_wpgatsby_page_path',
-							$page_path
-						);
-					}
-
-					if ( $modified ) {
-						update_post_meta(
-							$parent_id,
-							'_wpgatsby_node_modified',
-							$modified
-						);
-					}
-
-					if ( $remote_status ) {
-						update_post_meta(
-							$parent_id,
-							'_wpgatsby_node_remote_preview_status',
-							$remote_status
-						);
-					}
-
-					if ( $preview_context ) {
-						update_post_meta(
-							$parent_id,
-							'_wpgatsby_node_remote_preview_status_context',
-							$preview_context
-						);
-					}
-
-					// delete action monitor preview action.
-					// once we've saved this preview status as succes
-					// we don't need the preview action anymore.
-					$existing = new \WP_Query( [
-						'post_type'      => 'action_monitor',
-						'post_status'    => 'any',
-						'posts_per_page' => 1,
-						'no_found_rows'  => true,
-						'fields'         => 'ids',
-						'tax_query'      => [
-							'relation' => 'AND',
-							[
-								'taxonomy' => 'gatsby_action_ref_node_dbid',
-								'field'    => 'name',
-								'terms'    => sanitize_text_field( $parent_id ),
-							],
-							[
-								'taxonomy' => 'gatsby_action_stream_type',
-								'field'    => 'name',
-								'terms'    => 'PREVIEW',
-							]
-						],
-					] );
-
-					if ( isset( $existing->posts ) && ! empty( $existing->posts ) ) {
-						wp_delete_post( $existing->posts[0], true );
-					}
-
-					return [
-						'success' => true,
-					];
-				},
-			]
-		);
-
-		register_graphql_object_type(
-			'WPGatsbyPageNode',
-			[
-				'description' => __( 'A previewed Gatsby page node.' ),
-				'fields'      => [
-					'path' => [
-						'type' => 'String',
-					],
-				],
-			]
-		);
-
-		register_graphql_enum_type(
-			'WPGatsbyWPPreviewedNodeStatus',
-			[
-				'description' => __( 'The different statuses a Gatsby Preview can be in for a single node.', 'wp-gatsby' ),
-				'values'      => [
-					'NO_NODE_FOUND'                             => [
-						'value' => 'NO_NODE_FOUND',
-					],
-					'PREVIEW_READY'                             => [
-						'value' => 'PREVIEW_READY',
-					],
-					'REMOTE_NODE_NOT_YET_UPDATED'               => [
-						'value' => 'REMOTE_NODE_NOT_YET_UPDATED',
-					],
-					'NO_PREVIEW_PATH_FOUND'                     => [
-						'value' => 'NO_PREVIEW_PATH_FOUND',
-					],
-					'RECEIVED_PREVIEW_DATA_FROM_WRONG_URL'      => [
-						'value' => 'RECEIVED_PREVIEW_DATA_FROM_WRONG_URL',
-					],
-					'PREVIEW_PAGE_UPDATED_BUT_NOT_YET_DEPLOYED' => [
-						'value' => 'PREVIEW_PAGE_UPDATED_BUT_NOT_YET_DEPLOYED',
-					],
-				],
-			]
-		);
-
-		register_graphql_object_type(
-			'WPGatsbyPreviewStatus',
-			[
-				'description' => __( 'Check compatibility with a given version of gatsby-source-wordpress and the WordPress source site.' ),
-				'fields'      => [
-					'pageNode'       => [
-						'type' => 'WPGatsbyPageNode',
-					],
-					'statusType'     => [
-						'type' => 'WPGatsbyWPPreviewedNodeStatus',
-					],
-					'remoteStatus'   => [
-						'type' => 'WPGatsbyRemotePreviewStatusEnum',
-					],
-					'modifiedLocal'  => [
-						'type' => 'String',
-					],
-					'modifiedRemote' => [
-						'type' => 'String',
-					],
-					'statusContext'  => [
-						'type' => 'String',
-					],
-				],
-			]
-		);
-
-		register_graphql_field(
-			'WPGatsby',
-			'gatsbyPreviewStatus',
-			[
-				'description' => __( 'The current status of a Gatsby Preview.', 'wp-gatsby' ),
-				'type'        => 'WPGatsbyPreviewStatus',
-				'args'        => [
-					'nodeId' => [
-						'type'        => [ 'non_null' => 'Number' ],
-						'description' => __( 'The post id for the previewed node.', 'wp-gatsby' ),
-					],
-				],
-				'resolve'     => function( $root, $args, $context, $info ) {
-					$post_id = $args['nodeId'] ?? null;
-
-					// make sure post_id is a valid post
-					$post = get_post( $post_id );
-
-					$post_type_object = $post
-						? get_post_type_object( $post->post_type )
-						: null;
-
-					$user_can_edit_this_post = $post
-						? current_user_can(
-							$post_type_object->cap->edit_posts,
-							$post_id
-						)
-						: null;
-
-					if ( ! $post || ! $user_can_edit_this_post ) {
-						throw new UserError(
-							sprintf(
-								__(
-									'Sorry, you are not allowed to access the Preview status of post %1$s',
-									'wp-gatsby'
-								),
-								$post_id
-							)
-						);
-					}
-
-					if ( ! $post ) {
-						return [
-							'statusType' => 'NO_NODE_FOUND',
-						];
-					}
-
-					$found_preview_path_post_meta = get_post_meta(
-						$post_id,
-						'_wpgatsby_page_path',
-						true
-					);
-
-					$revision = Preview::getPreviewablePostObjectByPostId( $post_id );
-
-					$revision_modified = $revision->post_modified ?? null;
-
-					$modified = $revision_modified ?? $post->post_modified;
-
-					$gatsby_node_modified = get_post_meta(
-						$post_id,
-						'_wpgatsby_node_modified',
-						true
-					);
-
-					$remote_status = get_post_meta(
-						$post_id,
-						'_wpgatsby_node_remote_preview_status',
-						true
-					);
-
-					$node_modified_was_updated =
-						strtotime( $gatsby_node_modified ) >= strtotime( $modified );
-
-					if (
-						$node_modified_was_updated
-						&& (
-						'NO_PAGE_CREATED_FOR_PREVIEWED_NODE' === $remote_status
-						|| 'RECEIVED_PREVIEW_DATA_FROM_WRONG_URL' === $remote_status
-						)
-					) {
-						return [
-							'statusType'    => null,
-							'statusContext' => null,
-							'remoteStatus'  => $remote_status,
-						];
-					}
-
-					$node_was_updated = false;
-
-					if ( $node_modified_was_updated && $found_preview_path_post_meta ) {
-						$server_side = true;
-
-						$gatbsy_preview_frontend_url =
-							self::get_gatsby_preview_instance_url(
-								$server_side
-							);
-
-						$page_data_path = $found_preview_path_post_meta === "/" 
-							? "/index/"
-							: $found_preview_path_post_meta;
-		
-						$page_data_path_trimmed = trim( $page_data_path, "/" );
-
-						$modified_deployed_url =
-							$gatbsy_preview_frontend_url .
-							"page-data/$page_data_path_trimmed/page-data.json";
-
-						// check if node page was deployed
-						$request  = wp_remote_get( $modified_deployed_url );
-						$response = wp_remote_retrieve_body( $request );
-
-						$page_data = json_decode( $response );
-
-						$modified_response =
-							$page_data->result->pageContext->__wpGatsbyNodeModified
-							?? null;
-
-						error_log(print_r('$modified_response', true)); 
-						error_log(print_r($modified_response, true)); 
-						error_log(print_r('$modified', true)); 
-						error_log(print_r($modified, true)); 
-
-						$preview_was_deployed =
-							$modified_response &&
-							strtotime( $modified_response ) >= strtotime( $modified );
-
-						error_log(print_r('$preview_was_deployed', true)); 
-						error_log(print_r($preview_was_deployed, true)); 
-
-						if ( ! $preview_was_deployed ) {
-							// if preview was not yet deployed, send back PREVIEW_PAGE_UPDATED_BUT_NOT_YET_DEPLOYED.
-							return [
-								'statusType'    =>
-									'PREVIEW_PAGE_UPDATED_BUT_NOT_YET_DEPLOYED',
-								'statusContext' => null,
-								'remoteStatus'  => null,
-							];
-						} else {
-							// if it is deployed, send back PREVIEW_READY below.
-							$node_was_updated = true;
-						}
-					}
-
-					// if the node wasn't updated, then any status we have is stale.
-					$remote_status_type = $remote_status && $node_was_updated
-						? $remote_status
-						: null;
-
-					/**
-					 * We need the above check for wether the node was updated so we
-					 * don't show stale statuses on existing nodes, but in the case that
-					 * it's a brand new draft, $node_was_updated will always be false
-					 * because at this point we're potentially getting an error on a
-					 * node that was never created. So GATSBY_PREVIEW_PROCESS_ERROR is a
-					 * special case where we always need to show the status regardless
-					 * of wether the node was updated.
-					 */
-					if ( 'GATSBY_PREVIEW_PROCESS_ERROR' === $remote_status ) {
-						$remote_status_type = $remote_status;
-					}
-
-					$status_type = 'PREVIEW_READY';
-
-					if ( ! $node_was_updated ) {
-						$status_type = 'REMOTE_NODE_NOT_YET_UPDATED';
-					}
-
-					if ( ! $found_preview_path_post_meta ) {
-						$status_type = 'NO_PREVIEW_PATH_FOUND';
-					}
-
-					$status_context = get_post_meta(
-						$post_id,
-						'_wpgatsby_node_remote_preview_status_context',
-						true
-					);
-
-					if ( $status_context === '' ) {
-						$status_context = null;
-					}
-
-					$normalized_preview_page_path =
-						$found_preview_path_post_meta !== ''
-							? $found_preview_path_post_meta
-							: null;
-
-					return [
-						'statusType'     => $status_type,
-						'statusContext'  => $status_context,
-						'remoteStatus'   => $remote_status_type,
-						'pageNode'       => [
-							'path' => $normalized_preview_page_path,
-						],
-						'modifiedLocal'  => $modified,
-						'modifiedRemote' => $gatsby_node_modified,
-					];
-				},
-			]
-		);
-
-		register_graphql_field(
-			'WPGatsby',
-			'isPreviewFrontendOnline',
-			[
-				'description' => __( 'Wether or not the Preview frontend URL is online.', 'wp-gatsby' ),
-				'type'        => 'Boolean',
-				'resolve'     => function( $root, $args, $context, $info ) {
-					if ( ! is_user_logged_in() ) {
-						return false;
-					}
-
-					$preview_url = self::get_gatsby_preview_instance_url();
-
-					$request = wp_remote_get( $preview_url );
-
-					$request_was_successful =
-						$this->was_request_successful( $request );
-
-					return $request_was_successful;
-				},
-			]
-		);
-	}
-
 	/**
 	 * Determines wether or not the current global 
-	 * PHP context is related to Gatsby Preview
+	 * PHP context is related to Gatsby Content Sync Previews
 	 */
-	public static function is_gatsby_preview() {
-		$is_gatsby_preview 
+	public static function is_gatsby_content_sync_preview() {
+		$doing_graphql_request
+					= defined( 'GRAPHQL_REQUEST' ) && null !== GRAPHQL_REQUEST;
+
+		if ( $doing_graphql_request ) {
+			return false;
+		}
+
+		$enable_gatsby_preview = Settings::get_setting( 'enable_gatsby_preview' ) === 'on';
+
+		$is_gatsby_content_sync_preview 
 			= \is_preview()
 			|| isset( $_GET['preview_nonce'] )
 			|| isset( $_GET['gatsby_preview'] );
 
-		return $is_gatsby_preview;
+		return $is_gatsby_content_sync_preview && $enable_gatsby_preview;
 	}
 
 	/**
@@ -573,45 +68,21 @@ class PreviewMonitor extends Monitor {
 			$post = get_post( $_GET['preview_id'] );
 		}
 
-		// Determine the post_type of the post to be previewed
-		$post_type = $post->post_type ?? null;
+		if ( self::is_gatsby_content_sync_preview() && $post ) {
+			// Ensure the post_type is set to show_in_graphql
+			$post_type_object = $post->post_type ? get_post_type_object( $post->post_type ) : null;
 
-		// Ensure the post_type is set to show_in_graphql
-		$post_type_object = $post_type ? get_post_type_object( $post->post_type ) : null;
+			if ( $post_type_object && ! $post_type_object->show_in_graphql ?? true ) {
+				return plugin_dir_path( __FILE__ ) . '../../Admin/includes/post-type-not-shown-in-graphql.php';
+			}
 
-		if ( $post_type && ! $post_type_object->show_in_graphql ?? true ) {
-			return plugin_dir_path( __FILE__ ) . '../../Admin/includes/post-type-not-shown-in-graphql.php';
-		}
+			// WP doesn't call post_save for every second preview with no content changes.
+			// Since we're using post_save to trigger the webhook to Gatsby, we need to get WP to call post_save for this post.
+			do_action( 'save_post', $post->ID, $post, true );
 
-		// Determine if it's a preview
-		$is_gatsby_preview = self::is_gatsby_preview();
-		$is_wp_preview = is_preview();
-		$use_cloud_loader = self::get_setting( 'use_gatsby_content_sync' );
-
-		if ( $is_gatsby_preview && ! $is_wp_preview && $use_cloud_loader === "on" ) {
-			// WP doesn't call post_save for every second preview with no content changes. Since we're using post_save to trigger the webhook to Gatsby, we need to get WP to call post_save for this post.
-			$update = true;
-			$post_ID = $post->ID;
-			do_action( 'save_post', $post_ID, $post, $update );
-		}
-
-		if ( $is_gatsby_preview && $use_cloud_loader === "on" ) {
-			$post_ID = $post->ID;
-			$this->post_to_preview_instance( $post_ID, $post );
-		}
-
-		$preview_url = self::get_gatsby_preview_instance_url();
-		$enable_gatsby_preview = self::get_setting( 'enable_gatsby_preview' );
-
-		if ( 
-			$is_gatsby_preview && ( 
-				$preview_url || 
-					( ! $preview_url && $enable_gatsby_preview === 'on' ) 
-				)
-			) {
+			$this->post_to_preview_instance( $post->ID, $post );
+		
 			return trailingslashit( dirname( __FILE__ ) ) . '../../Admin/includes/preview-template.php';
-		} elseif ( $is_gatsby_preview && ! $preview_url ) {
-			return trailingslashit( dirname( __FILE__ ) ) . '../../Admin/includes/no-preview-url-set.php';
 		}
 
 		return $template;
@@ -622,13 +93,13 @@ class PreviewMonitor extends Monitor {
 	 */
 	public function post_to_preview_instance( $post_ID, $post ) {
 		$revisions_are_disabled = 
-			!wp_revisions_enabled( $post );
+			! wp_revisions_enabled( $post );
 
 		if (
 			defined( 'DOING_AUTOSAVE' )
 			&& DOING_AUTOSAVE
 			// if revisions are disabled, our autosave is our preview
-			&& !$revisions_are_disabled
+			&& ! $revisions_are_disabled
 		) {
 			return;
 		}
@@ -653,18 +124,17 @@ class PreviewMonitor extends Monitor {
 		$is_revision = $post->post_type === 'revision';
 		$is_draft = $post->post_status === 'draft';
 
-		$is_gatsby_preview = self::is_gatsby_preview();
+		$is_gatsby_content_sync_preview = self::is_gatsby_content_sync_preview();
 
 
 		if (
 			! $is_draft
 			&& ! $is_revision
 			&& ! $is_new_post_draft
-			&& ! $is_gatsby_preview
+			&& ! $is_gatsby_content_sync_preview
 		) {
 			return;
 		}
-
 
 		$token = \WPGatsby\GraphQL\Auth::get_token();
 
@@ -679,21 +149,16 @@ class PreviewMonitor extends Monitor {
 
 		$original_post = get_post( $post->post_parent );
 
-		$use_cloud_loader = self::get_setting( 'use_gatsby_content_sync' );
-
 		$this_is_a_publish_not_a_preview = 
 			$original_post
 			&& $original_post->post_modified === $post->post_modified
-			&& ( $use_cloud_loader !== "on" || ! $is_gatsby_preview );
+			&& ! $is_gatsby_content_sync_preview;
 
 			
 		if ( $this_is_a_publish_not_a_preview ) {
 			// we will handle this in ActionMonitor.php, not here.
 			return;
 		}
-
-
-		$parent_post_id = $original_post->ID ?? $post_ID;
 
 		$post_type_object = $original_post
 			? \get_post_type_object( $original_post->post_type )
@@ -705,14 +170,7 @@ class PreviewMonitor extends Monitor {
 			return;
 		}
 		
-
-
-		// otherwise store this modified time so we can compare it next time.
-		update_post_meta(
-			$parent_post_id,
-			self::$last_sent_modified_time_key,
-			$post->post_modified
-		);
+		$parent_post_id = $original_post->ID ?? $post_ID;
 
 		$global_relay_id = Relay::toGlobalId(
 			'post',
@@ -784,89 +242,18 @@ class PreviewMonitor extends Monitor {
 			]
 		);
 
-		// this is used to optimistically load the preview iframe
-		// we also check if the frontend is responding to requests from the
-		// preview template JS.
-		$webhook_success = $this::was_request_successful( $response );
-
-		update_option(
-			'_wp_gatsby_preview_webhook_is_online',
-			$webhook_success, // boolean.
-			true
-		);
-
-		if ( ! $webhook_success ) {
-			error_log(
-				'WPGatsby couldn\'t reach the Preview webhook set in plugin options.'
-			);
+		if ( \is_wp_error( $response ) ) {
+			error_log( "WPGatsby couldn\'t POST to the Preview webhook set in plugin options.\nWebhook returned error: {$response->get_error_message()}" );
 		}
-	}
-
-
-	/**
-	 * This is used to print out the client JS file directly to the
-	 * Preview template html
-	 */
-	public static function printFileContents( $fileName ) {
-		$pluginDirectory = plugin_dir_path( __FILE__ );
-		$filePath        = $pluginDirectory . $fileName;
-		echo file_get_contents( $filePath );
-	}
-
-	/**
-	 * Get a WPGatsby setting by setting key
-	 *
-	 * @param string $key The name of the setting to get the value for
-	 */
-	static function get_setting( string $key ) {
-		$wpgatsby_settings = get_option( 'wpgatsby_settings' );
-
-		return isset( $wpgatsby_settings[ $key ] ) ? $wpgatsby_settings[ $key ] : null;
-	}
-
-	/**
-	 * Get the normalized/validated frontend url of the Gatsby Preview
-	 */
-	static function get_gatsby_preview_instance_url( $server_side = false ) {
-		$preview_url = self::get_setting( 'preview_instance_url' );
-
-		$preview_url_exploded = explode( ',', $preview_url );
-
-		// this allows using a different url as the frontend url
-		// on the server side and in the preview browser
-		// for our tests we need to use http://host.docker.internal:8000 for the PHP
-		// side checks for wether or not the page-data.json has deployed
-		// but we need to run the preview-template.php code with the frontend url as
-		// http://localhost:8000
-		// So this allows passing
-		// "http://host.docker.internal:8000,http://localhost:8000"
-		// as the server preview frontend and template preview frontend
-		if ( count( $preview_url_exploded ) > 1 ) {
-			$preview_url = $server_side
-				? $preview_url_exploded[0]
-				: $preview_url_exploded[1];
-		}
-
-		if ( ! $preview_url || ! filter_var( $preview_url, FILTER_VALIDATE_URL ) ) {
-			return false;
-		}
-
-		if ( substr( $preview_url, -1 ) !== '/' ) {
-			$preview_url .= '/';
-		}
-
-		return $preview_url;
 	}
 
 	/**
 	 * Get the Gatsby Preview instance refresh webhook
 	 */
 	static function get_gatsby_preview_webhook() {
-		$preview_webhook = self::get_setting( 'preview_api_webhook' );
+		$preview_webhook = Settings::get_setting( 'preview_api_webhook' );
 
-		if ( ! $preview_webhook
-				|| ! filter_var( $preview_webhook, FILTER_VALIDATE_URL )
-		) {
+		if ( ! $preview_webhook || ! filter_var( $preview_webhook, FILTER_VALIDATE_URL ) ) {
 			return false;
 		}
 
@@ -876,29 +263,4 @@ class PreviewMonitor extends Monitor {
 
 		return $preview_webhook;
 	}
-
-	public static function was_request_successful( $response ) {
-		$is_wp_error = is_wp_error( $response );
-
-		$status_code             = ! $is_wp_error ? $response['response']['code'] ?? null : null;
-		$response_message_was_ok = ! $is_wp_error ? $response['response']['message'] === 'OK' ?? null : null;
-
-		// this is used to optimistically load the preview iframe
-		// we also check if the frontend is responding to requests from the
-		// preview template JS
-		$success =
-			! $is_wp_error &&
-			(
-				$status_code === 200 ||
-				$status_code === 204 ||
-				$response_message_was_ok
-			);
-
-		if ( ! $success ) {
-			error_log($status_code);
-		}
-
-		return $success;
-	}
-	
 }
