+++ src/Admin/Settings.php
@@ -22,8 +22,7 @@ class Settings {
 	/**
 	 * If the settings haven't been saved yet, save the JWT once to prevent it from re-generating.
 	 */
-	public function set_default_jwt_key() {
-
+	public function set_default_jwt_key() {	
 		// Get the JWT Secret
 		$default_secret = self::get_setting( 'preview_jwt_secret' );
 
@@ -211,29 +210,10 @@ class Settings {
 		$settings_fields = [
 			'wpgatsby_settings' => [
 				[
-					'name'              => 'builds_api_webhook',
-					'label'             => __( 'Builds Webhook URL', 'wpgatsby_settings' ),
-					'desc'              => __( 'Use a comma-separated list to configure multiple webhooks.', 'wpgatsby_settings' ),
-					'placeholder'       => __( 'https://', 'wpgatsby_settings' ),
-					'type'              => 'text',
-					'sanitize_callback' => function ( $input ) {
-						return $this->sanitize_url_field( $input );
-					},
-				],
-				[
 					'name'  => 'enable_gatsby_preview',
 					'label' => __( 'Enable Gatsby Preview?', 'wpgatsby_settings' ),
 					'desc'  => __( 'Yes, allow Gatsby to take over WordPress previews.', 'wpgatsby_settings' ),
 					'type'  => 'checkbox',
-				],
-				[
-					'name'              => 'preview_instance_url',
-					'label'             => __( 'Preview Frontend URL', 'wpgatsby_settings' ),
-					'placeholder'       => __( 'https://', 'wpgatsby_settings' ),
-					'type'              => 'text',
-					'sanitize_callback' => function ( $input ) {
-						return $this->sanitize_url_field( $input );
-					},
 				],
 				[
 					'name'              => 'preview_api_webhook',
@@ -246,10 +226,14 @@ class Settings {
 					},
 				],
 				[
-					'name'  => 'use_gatsby_content_sync',
-					'label' => __( 'Use Gatsby Content Sync? (BETA feature)', 'wpgatsby_settings' ),
-					'desc'  => __( 'Yes, I want to use Gatsby Cloud to redirect admins to the right preview page. (BETA but recommended)', 'wpgatsby_settings' ),
-					'type'  => 'checkbox',
+					'name'              => 'builds_api_webhook',
+					'label'             => __( 'Builds Webhook URL', 'wpgatsby_settings' ),
+					'desc'              => __( 'Use a comma-separated list to configure multiple webhooks.', 'wpgatsby_settings' ),
+					'placeholder'       => __( 'https://', 'wpgatsby_settings' ),
+					'type'              => 'text',
+					'sanitize_callback' => function ( $input ) {
+						return $this->sanitize_url_field( $input );
+					},
 				],
 				[
 					'name'              => 'gatsby_content_sync_url',
