--- src/Admin/assets/.browserslistrc
@@ -1,3 +1,0 @@
->0.2%
-not dead
-IE 11
