--- src/Admin/assets/postcss.config.js
@@ -1,10 +1,0 @@
-const postcssPresetEnv = require("postcss-preset-env")
-
-module.exports = {
-	plugins: [
-		postcssPresetEnv(),
-		require("cssnano")({
-			preset: "default",
-		}),
-	],
-}
