+++ src/GraphQL/Auth.php
@@ -3,20 +3,19 @@
 namespace WPGatsby\GraphQL;
 
 use \Firebase\JWT\JWT;
-use \WPGatsby\Admin\Preview;
+use \WPGatsby\Admin\Settings;
 
 class Auth {
 	static function get_token() {
 		$site_url    = get_bloginfo( 'url' );
-		$preview_url = Preview::get_gatsby_preview_instance_url();
-		$secret      = Preview::get_setting( 'preview_jwt_secret' );
+		$secret      = Settings::get_setting( 'preview_jwt_secret' );
 		$now         = time();
 		$expiry      = $now + 3600;
 		$user_id     = get_current_user_id();
 
 		$payload = [
 			'iss'  => $site_url,
-			'aud'  => $preview_url,
+			'aud'  => $site_url,
 			'iat'  => $now,
 			'nbf'  => $now,
 			'exp'  => $expiry,
